+++ examples/start-electron-app/src/electron/app.ts
@@ -1,21 +1,22 @@
-import { Autowired, IBrowserWindow, Init, Module } from '@electron-boot/framework';
+import { Autowired, Init, Module } from '@electron-boot/framework';
 import { LoggerFactory } from '@electron-boot/logger';
 import { SystemController } from './controller/system.controller';
 import defaultConfig from './config/config.default';
+import { UpdaterModule } from '@electron-boot/updater';
+import MainWindow from './windows/main/main.window';
 
 @Module({
-  providers: [SystemController],
+  imports: [UpdaterModule],
+  providers: [SystemController, MainWindow],
   configs: [{ default: defaultConfig }],
 })
 export class AppModule {
   private logger = LoggerFactory.getLogger(AppModule);
 
   @Autowired()
-  mainWindow: IBrowserWindow;
+  mainWindow: MainWindow;
   @Init()
   async init() {
-    this.mainWindow.on('ready-to-show', () => {
-      this.mainWindow.show();
-    });
+    console.log('进来了');
   }
 }