+++ examples/start-electron-app/src/electron/controller/system.controller.ts
@@ -1,5 +1,7 @@
 import { Event, Controller, Autowired, Context } from '@electron-boot/framework';
 import { LoggerFactory } from '@electron-boot/logger';
+import { IpcMainEvent } from 'electron';
+import { UpdaterService } from '@electron-boot/updater';
 
 @Controller
 export class SystemController {
@@ -8,12 +10,22 @@ export class SystemController {
   @Autowired()
   ctx: Context;
 
+  @Autowired()
+  event: IpcMainEvent;
+
+  @Autowired()
+  updaterService: UpdaterService;
   @Event
   async showMessage(data: string) {
+    console.log(this.event);
     function sleep(time) {
       return new Promise(resolve => setTimeout(resolve, time));
     }
     await sleep(5000);
     return data + 'world';
+  }
+  @Event()
+  async checkUpdate() {
+    await this.updaterService.checkUpdate();
   }
 }