+++ packages/framework/src/supports/index.ts
@@ -1,6 +1,4 @@
-export * from './client/window.client';
-export * from './factory/abstract.service.factory';
-export * from './factory/window.service.factory';
+export * from './service/abstract.factory.service';
 export * from './service/aspect.service';
 export * from './service/config.service';
 export * from './service/decorator.service';