+++ packages/updater/src/updater.ts
@@ -1,10 +1,16 @@
 import { Autowired, IApplicationContext, Init, Module } from '@electron-boot/framework';
 import { UpdaterService } from './updater.service';
+import { IBrowserWindow } from '@electron-boot/framework/src';
 
-@Module()
+@Module({
+  providers: [UpdaterService],
+})
 export class UpdaterModule {
-  @Autowired()
+  @Autowired('rootContext')
   applicationContext: IApplicationContext;
+
+  @Autowired()
+  private mainWindow: IBrowserWindow;
   @Init()
   async init() {
     await this.applicationContext.getAsync(UpdaterService);