+++ examples/start-electron-app/CHANGELOG.md
@@ -1,5 +1,35 @@
 # start-electron-app
 
+## 1.0.0
+
+### Major Changes
+
+- a74f6dd: change typescript version >= 5.2
+- 17e942c: fix: get request ctx
+
+### Minor Changes
+
+- 17e942c: perf: Streamline Code
+
+### Patch Changes
+
+- Updated dependencies [a74f6dd]
+- Updated dependencies [17e942c]
+- Updated dependencies
+- Updated dependencies [17e942c]
+  - @electron-boot/framework@1.0.0
+  - @electron-boot/updater@1.0.0
+  - @electron-boot/logger@1.0.0
+
+## 1.0.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @electron-boot/framework@1.0.0-beta.8
+  - @electron-boot/updater@1.0.0-beta.8
+  - @electron-boot/logger@1.0.0-beta.8
+
 ## 1.0.0-beta.7
 
 ### Major Changes