+++ packages/updater/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @electron-boot/updater
 
+## 1.0.0
+
+### Major Changes
+
+- a74f6dd: change typescript version >= 5.2
+- fix: change typescript to 5.2.2 and fix bug
+- 17e942c: fix: get request ctx
+
+### Minor Changes
+
+- 17e942c: perf: Streamline Code
+
+### Patch Changes
+
+- Updated dependencies [a74f6dd]
+- Updated dependencies [17e942c]
+- Updated dependencies
+- Updated dependencies [17e942c]
+  - @electron-boot/framework@1.0.0
+  - @electron-boot/logger@1.0.0
+
+## 1.0.0-beta.8
+
+### Major Changes
+
+- fix: change typescript to 5.2.2 and fix bug
+
+### Patch Changes
+
+- Updated dependencies
+  - @electron-boot/framework@1.0.0-beta.8
+  - @electron-boot/logger@1.0.0-beta.8
+
 ## 1.0.0-beta.7
 
 ### Major Changes