+++ packages/vite-plugin-electron/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @electron-boot/vite-plugin-electron
 
+## 1.0.0
+
+### Major Changes
+
+- a74f6dd: change typescript version >= 5.2
+- fix: change typescript to 5.2.2 and fix bug
+- 17e942c: fix: get request ctx
+
+### Minor Changes
+
+- 17e942c: perf: Streamline Code
+
+## 1.0.0-beta.8
+
+### Major Changes
+
+- fix: change typescript to 5.2.2 and fix bug
+
 ## 1.0.0-beta.7
 
 ### Major Changes